fix: スクレイピングの関数を一つに統一
